SOFT: Fixed clamping distance bugs.

+++ src/gui/myframe.h
@@ -319,7 +319,7 @@ class MyFrame : public MyFrame_Base, public UpdatedValuesReceiver
     long m_CurrentForce;												/**< Current force */
     int m_CurrentForceUpdateDelay;							/**< Counting variable that the force values is not updated always in the GUI. */
     wxString m_ForceUnit;												/**< Current force unit (N or kPa) */
-    long m_ClampingDistance;										/**< Clamping distance */
+    double m_ClampingDistance;									/**< Clamping distance */
     long m_PreloadDistance;											/**< Preload distance */
 
     wxDECLARE_EVENT_TABLE();
